+++ SampleMetadata/Sample.cs
@@ -1,0 +1,13 @@
+using System;
+using System.Reflection;
+
+[assembly: AssemblyMetadata("Custom", "Custom")]
+[assembly: CustomMetadata("Custom", "Custom")]
+
+namespace SampleMetadata
+{
+    public class Sample
+    {
+
+    }
+}