+++ SampleMetadataAttribute/CustomMetadataAttribute.cs
@@ -1,0 +1,10 @@
+using System;
+using System.Reflection;
+
+[assembly: AssemblyMetadata("Custom", "Custom")]
+
+[AttributeUsage(AttributeTargets.Assembly)]
+public class CustomMetadataAttribute : Attribute
+{
+    public CustomMetadataAttribute(string key, string value) { }
+}